Added tracking page (#7)
* Added Order Tracking feature and Input Mask

* Add TrackBody component and Slide-in animations

This commit introduces a new component: TrackBody, which includes order status, estimated delivery time and a map with markers. Slide-in animations were also added to tailwind.config.ts to provide a smoother user interface experience.

In addition, dependencies were added to package-lock.json (e.g., pigeon-maps) to support the new functionalities.
This change was necessary to provide live tracking and to better visualize the order's progress.

* Fix track image import path in TrackSearchCard

This commit changes the import path for the Image component in TrackSearchCard.tsx. Instead of importing Image from 'next/image', it now imports Image from 'next/legacy/image'.

This change is necessary for backward compatibility reasons as the newer version of 'next/image' might not support some older features used within the TrackSearchCard. The switch to 'next/legacy/image' ensures that these features are supported and prevent potential issues with image rendering in the TrackSearchCard component.

* Add dialog module and update TrackBody tracking link